+++ test/external/kovan/Aave_NFT_Timeout_Credit.js
@@ -96,6 +96,7 @@ contract('End to End NFT Collateral Scenario (timeout)', (accounts) => {
 
         await conditionStoreManager.initialize(
             agreementStoreManager.address,
+            owner,
             { from: deployer }
         )
